Add files via upload
Classification Task- To classify returns on the basis customer and deklivery KPIs for fashion e-commerce company in Germany
The problem in hand was that the company was getting over 50% of product returns. Used Random Forest and Support Vector Machine algorithm to classify them. Achieved an accuracy of 71% and Precision of 83% (for return class) with the most important features being delivery lead time, gender and cloth type. This would help company to work on the delivery lead times in the regions where returns are more.